refactor: make sortString available for all types
