+++ types/Color.php
@@ -68,4 +68,45 @@ class Color extends AbstractBaseType {
         return "$html";
     }
 
+    /**
+     * Sort by the hue of a color, not by its hex-representation
+     */
+    public function getSortString($value) {
+        $hue = $this->getHue(parent::getSortString($value));
+        return $hue;
+    }
+
+    /**
+     * Calculate the hue of a color to use it for sorting so we can sort similar colors together.
+     *
+     * @param string $color the color as #RRGGBB
+     * @return float|int
+     */
+    protected function getHue($color) {
+        if (!preg_match('/^#[0-9A-F]{6}$/i', $color)) {
+            return 0;
+        }
+
+        $red   = hexdec(substr($color, 1, 2));
+        $green = hexdec(substr($color, 3, 2));
+        $blue  = hexdec(substr($color, 5, 2));
+
+        $min = min([$red, $green, $blue]);
+        $max = max([$red, $green, $blue]);
+
+        if ($max == $red) {
+            $hue = ($green-$blue)/($max-$min);
+        }
+        if ($max == $green) {
+            $hue = 2 + ($blue-$red)/($max-$min);
+        }
+        if ($max == $blue) {
+            $hue = 4 + ($red-$green)/($max-$min);
+        }
+        $hue = $hue * 60;
+        if ($hue < 0) {
+            $hue += 360;
+        }
+        return $hue;
+    }
 }
